+++ libs/providers/modal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @langchain/modal
 
+## 0.1.2
+
+### Patch Changes
+
+- [#197](https://github.com/langchain-ai/deepagentsjs/pull/197) [`e4b5892`](https://github.com/langchain-ai/deepagentsjs/commit/e4b5892b0e171cf33b75c8e2c93665ce97f87638) Thanks [@christian-bromann](https://github.com/christian-bromann)! - fix(deepagents): runtime agnostic sandbox operations
+
+## 0.1.1
+
+### Patch Changes
+
+- [#194](https://github.com/langchain-ai/deepagentsjs/pull/194) [`731b01e`](https://github.com/langchain-ai/deepagentsjs/commit/731b01ed172dd4cbc0fa45f0189723ad6890f366) Thanks [@christian-bromann](https://github.com/christian-bromann)! - fix(deepagents): polish sandbox interfaces
+
 ## 0.1.0
 
 ### Minor Changes